build: fix Windows build failures
- Force LF endings on JUCE checkout via `GIT_CONFIG core.autocrlf=false`
  and on the patch file via `.gitattributes`, so the VST3 host patch
  applies cleanly on Windows where Git defaults to CRLF
- Normalize `juce-vst3-param-cache-resync.patch` to pure LF (was mixed
  CRLF/LF, which made `patch.exe` fail with "different line endings")
- Replace `std::from_chars` with `scn::scan_value<float>` in
  `FaustPlugin` metadata parsing and `std::strtod` in the test CLAP
  plugins, fixing compile errors where `from_chars` for floats is
  unavailable
- Add `static_cast<uint32_t>` on `CLAP_NOTE_DIALECT_MIDI` to fix a
  sign-conversion compile error in
  `ClapPlugin::create_ports_from_clap_plugin`
- Update `REUSE.toml` license annotation path from `generated-juce` to
  `generated-cpp` to match the new directory layout

+++ src/plugins/clap_plugin.cpp
@@ -1416,7 +1416,7 @@ ClapPlugin::create_ports_from_clap_plugin ()
           pimpl_->note_in_supported_dialects_.push_back (
             pimpl_->plugin_->notePortsGet (i, true, &info)
               ? info.supported_dialects
-              : CLAP_NOTE_DIALECT_MIDI);
+              : static_cast<uint32_t> (CLAP_NOTE_DIALECT_MIDI));
           auto port_ref = utils::create_object<dsp::MidiPort> (
             registry (),
             utils::Utf8String::from_utf8_encoded_string (
